// src/app/models/associate.ts
export class Associate {
    id: number;
    idTxt: string;
    name: string;
    email: string;
    mobile: string;
    picture: string;
    pictureBase64String: string;
    status: string;
    level: number;
    levelTxt: string;
    remark: string;
    strength: string;
    weakness: string;
    gender: string;
    isFresher: boolean;
    otherSkill: string;
    strongSkills: string;    
}